review: Reduce duplication in header menu items by moving the items into a separate component
